confirma overwrite quando diretório-destino já existe

+++ instalar.py
@@ -1,4 +1,5 @@
 # -*- encoding: utf-8 -*-
+from __future__ import print_function
 import os.path
 import shutil
 import sys
@@ -33,7 +34,17 @@ parser.add_argument("-v, --verbose", help="Exibe informações adicionais", dest
 args = parser.parse_args()
 
 if os.path.exists(DIRETORIO_DE_INSTALACAO):
-    shutil.rmtree(DIRETORIO_DE_INSTALACAO)
+    opt = ""
+
+    while opt.lower() not in ["s", "n"]:
+        print("Diretório {0} existe; remover?".format(DIRETORIO_DE_INSTALACAO), end=" ")
+        opt = raw_input("[s/n] ")
+
+    if opt == "s":
+        shutil.rmtree(DIRETORIO_DE_INSTALACAO)
+    else:
+        print("Instalação cancelada")
+        exit(0)
 else:
     if args.verbose:
         print("Diretório {0} não existe; criando...".format(DIRETORIO_DE_INSTALACAO))
